backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Steve Dignam <steve@dignam.xyz>", "Christopher Dignam <chris@dignam.xyz>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.12.4"
Django = "^3.2.9"
dj-database-url = "0.4.2"
gunicorn = "20.1.0"
pyyaml = "^5.1"
django-user-sessions = "^1.6"
python-dotenv = "0.21.0"
typing_extensions = "^4.8.0"
icalendar = "^4.0"
djangorestframework-types = "^0.2.0"
orjson = "3.8.1"
psycopg2-binary = "^2.9.3"
boto3 = "^1.24.44"
pydantic = {extras = ["email"], version = "^1.10.2"}
yarl = "^1.8.1"
advocate = "^1.0.0"
Pillow = "^9.3.0"
httpx = "^0.23.1"
asyncpg = "^0.27.0"
asyncpg-stubs = "^0.27.0"
typer = "^0.7.0"
structlog = "^22.3.0"
types-Pillow = "^9.3.0"
pillow-heif = "^0.9.0"
tldextract = "^3.4.0"
sentry-sdk = "^1.35.0"
ably = "^2.0.1"
recipe-scrapers = "^14.52.0"

[tool.poetry.dev-dependencies]
pytest = "7.2.0"
pytest-django = "4.5.2"
ipdb = "^0.13.9"
django-types = "^0.19.1"
types-PyYAML = "^6.0.0"
types-requests = "^2.25.11"
syrupy = "^3.0.5"
boto3-stubs = {extras = ["s3"], version = "^1.24.44"}
types-urllib3 = "^1.26.23"
types-dj-database-url = "^1.0.0"
mypy = "^1.7.0"
ruff = "^0.1.6"

[tool.ruff]
select = [
    "E",
    "F",
    "TID252",
    "I001",
    "T20",
    "C4",
    "UP",
    "N",
    "BLE",
    "B",
    "RET",
    "SIM",
    "ARG",
    "DTZ",
    "ERA",
    "RUF",
]
src = ["recipeyak"]
ignore = [
    "E501", # line length is handled by black
    "ARG001", # pytest fixtures mess with this
    "ARG002", # sometimes parent classes require params
    "N806", # django migrations violate this a lot
    "B008", # type.Argument violates this
    "N815", # we mix and match casing in serializers
    "ARG005", # monkey patching false positive with *args, and **kwargs
    "N805", # false positive with pydantic
    "N802", # DRF serializers mess with this
    "RET505", # if, elif, else are fine with returns
    "RET507", # if, elif, else with continue is also fine
    "RET508", # if, elif, else with break is also fine
]
unfixable = ["ERA001"] # we want to check ourselves before we delete commented out code. Also this doesn't cleanup all of it.
target-version = "py311"

[tool.ruff.lint.per-file-ignores]
"recipeyak/migrations/*" = [
    # it's very common to use class vars in Django migrations.
    "RUF012",
]

[tool.ruff.isort]
known-third-party = ["django"]
known-first-party = ["recipeyak"]

[tool.ruff.flake8-tidy-imports]
# Disallow all relative imports.
ban-relative-imports = "all"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "recipeyak.django.settings"
python_files = "*_test.py *_tests.py"
addopts = "-s --pdbcls IPython.terminal.debugger:TerminalPdb --reuse-db -p no:warnings"

[tool.mypy]
mypy_path = "./typings"
files = ["**/*.py"]

show_column_numbers = true
show_error_codes = true

disallow_any_unimported=false
disallow_any_expr=false
disallow_any_decorated=false
disallow_any_explicit=false
disallow_any_generics=true
disallow_subclassing_any=false

disallow_untyped_calls=false
disallow_untyped_defs=true
disallow_incomplete_defs=true
check_untyped_defs=true
disallow_untyped_decorators=true

no_implicit_optional=true
strict_optional=true

warn_redundant_casts=true
warn_unused_ignores=true
warn_no_return=true
warn_return_any=true
# doesn't consider how methods can change the value of properties
warn_unreachable=false

strict_equality=true

ignore_missing_imports=false

plugins = [
  "pydantic.mypy"
]

[[tool.mypy.overrides]]
module = [
    'icalendar.*',
    'advocate.*',
    'user_sessions.*',
    'ably.*'
]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true
